+++ frontend/core/containers/thread/DashboardThread/logic/useGaussBlur.ts
